Rename cairo_types to types
