+++ src/common/mod.rs
@@ -1,4 +1,4 @@
-pub mod cairo_types;
 pub mod errors;
 pub mod starknet_constants;
 pub mod traits;
+pub mod types;
